+++ src/004_nand_hdl.html.pm
@@ -4,8 +4,8 @@
 
 ◊(yt "RQzmQvi1tdU")
 
-A NAND gate is made up of 1 AND gate, and 1 NOT gate to invert its output. It only outputs a 0 when both inputs are 1. The truth table looks like this:
-
+◊slide{
+◊h2{The NAND Gate}
 ◊(table 
     (table-headings (list "A" "B" "OUT"))
 (table-entries 
@@ -14,35 +14,75 @@ A NAND gate is made up of 1 AND gate, and 1 NOT gate to invert its output. It on
 (list "1" "0" "1")
 (list "1" "1" "0")
 ) #:caption (list "NAND" ◊img[#:src "/images/nand_gate.svg"]{}))
+}
+Above is the ◊strong{NAND} gate(aka ◊strong{NOT AND}), which outputs ◊code{0} ◊strong{if and only if} both inputs are ◊code{1}, and outputs ◊code{1} otherwise. It looks like an ◊strong{AND} gate, just with a ◊strong{bubble} at the end, which represents the ◊strong{NOT} operation.
 
-It's known as the universal gate because it is the building block for all the other gates, like AND, NOT, OR, etc
-It's so popular in fact, that it gets its own special little notation/symbol to shorten drawing it out, notice a dot at the end of an AND gate.
+◊slide{
+◊h2{Building the NAND Gate}
+◊img[#:src "/images/relay.gif"]{}
+◊h3{◊a[#:href "http://nandgame.com/"]{Let's build the NAND gate!}}
+}
+We can build a ◊strong{NAND} gate by cleverly arranging Relays. Relay's have a wire wrapped around an magnetic rod. When electricity runs through the coil, it creates a magnetic field that can pull down a lever which then outputs electricity depending.
 
-I know, it's a little weird, because as I just said, NAND is composed of two gates, 1 AND, 1 NOT, so why do we have to build them from a NAND gate if we already have them? Isn't it more efficent to just use the AND and NOT gates from the get go, rather than having to build them through a NAND gate? Why do we have to go through the middle man of the NAND gate?
+◊slide{
+◊h2{The Universal Gate}
+The ◊strong{NAND} is also known as the ◊em{"universal gate"}, meaning it can be used to build all the other gates, like ◊em{AND, NOT, OR, etc}.
+}
 
-This is an electrical engineering question that involves voltages and transistors and etc, that I'm not qualified to answer, but what I can say is that its not that efficent to cut out the NAND gate, sometimes it makes things worse, and I will link to sources below if you're curious.
+◊slide{
+◊h3{Elementary Logic Gates}
+◊ol{
+    ◊li{Not}
+    ◊li{And}
+    ◊li{Or}
+    ◊li{Xor}
+    ◊li{Mux(hardest), aka selector in nandgame}
+    ◊li{DMux(hardest), aka switch in nandgame}
+}
 
-For this course in particular, we're just going to focus on using the abstraction of the truth table and combing gates together, not worrying about where the logic gates come from, how they are implemented, or how efficent things are. This is the basis of "abstraction", no one needs to know mechanical engineering to drive a car, instead an interface is exposed, the steering wheel and the pedal, to just tell you what does what, so you don't need to be concerned with how every gear turns. If you're curious about the HOW, look into physics and EE.
+◊h3{16-bit variants}
+◊em{nandgame does not have these but it's easy to do in HDL}
+◊ol{
+    ◊li{Not16}
+    ◊li{And16}
+    ◊li{Or16}
+    ◊li{Mux16}
+}
 
-Alright, you are to start off with a NAND gate, and work your way up to build other gates, and when you have built those other gates, you can use those to build other gates. Let's hop to the hardware simulator so you get what I mean.
+◊h3{Multiway Variants}
+◊ol{
+    ◊li{Or8Way}
+    ◊li{Mux4Way16}
+    ◊li{Mux8Way16}
+    ◊li{DMux4Way}
+    ◊li{DMux8Way}
+}
+}
 
-The hardware simulator runs on mac windows and linux, but we have to install java, so go to this link: https://www.java.com/en/download/
-click next a bunch of times.
+So, you are tasked to start off with a NAND gate, and work your way up to build other gates, and when you have built those other gates, you can use those gates you built to build other gates. It's recommended you do it in the order above because the difficulty increases as you go down the list.
 
-extract the zip file of the hardware simulator and run it
+◊slide{
+◊h2{Resources}
+◊h3{◊a[#:href "https://www.nandgame.com/"]{Nand Game} (drag and drop gates)}
+◊h3{◊a[#:href "https://nand2tetris.github.io/web-ide/chip/"]{Nand2Tetris Web IDE} (text based chip diagrams)}
+I recommend you ◊strong{first} use the visual NAND game, since the UI is more intuitive, and then transcribe what you did to HDL using the online web IDE.
 
-visualization tools:
-http://nandgame.com/
-circutverse
+◊h3{◊a[#:href "/media/project-1.pdf"]{Chip truth tables and specs} (pages 19 to 28)}
+◊; explain the test script, compare file, etc
+}
 
-HDL:
-No loops, no conditional(if else), just plugin in variables. Case sensitive, madatory semicolons, commas, be careful for typos
+◊slide{
+◊h2{HDL}
+HDL also known as ◊em{Hardware Description Language} is what real chip designers use and we have stripped it down to its bare bones to make it easier to learn.
 
-internal pins
+It is ◊strong{NOT} a programming language in the typical sense.
+◊ul{
+    ◊li{No loops}
+    ◊li{No conditionals(if else)}
+    ◊li{No Functions}
+}
+You just plug in variables to outputs.
+It is ◊strong{case sensitive}, mandatory semicolons, commands, becareful of typos!
+}
 
-test script(you're really never going to touch the test script)
-compare file
-
-I recommend getting things working with visual tool first, then transcribe into text
-
- https://electronics.stackexchange.com/questions/394890/why-are-nand-gates-used-to-make-and-gates-in-computers
+◊; internal pins